Fix trailing character for phone number input mask
Because of issues relating to textMask and ionInput, an issue was causing a trailing number to appear at the end of a phone number. For now, I just have a hacky, hardcoded fix put in, but because this is seemingly the only place in the entire app where I will need input masking, I'm mildly okay with that.