Fix filtering after query
Most of the issues I was running into seemingly had to do with not having the geohash be indexed in my firebase rules, so while this commit does not reflect that, that was the real issue. Note to everyone who uses the GeoFire library (or does any amount of server-side filtering through their queries), index your data!